refactored name on model package to core as it now contains exceptions as well
