+++ bot/core/entity/LoggedInUser.py
