+++ bot/core/errors/Exceptions.py
