+++ bot/core/errors/__init__.py
